Major updates
1. Updated api/contact.js to push contact requests to firestore db and added mail notification to me in addition to mail acknowledgment to requestor.
2. Moved contact form to its own Component class.
3. Included link to Netlify CMS Link
4. Fixed Netlify CMS content Path
5. Renamed writing content folder name to "articles"
