+++ web/package.json
@@ -28,7 +28,7 @@
     "dateformat": "^3.0.3",
     "gray-matter": "^4.0.2",
     "isomorphic-unfetch": "^3.0.0",
-    "next": "^8.1.0",
+    "next": "^9.0.0",
     "raw-loader": "^3.0.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
@@ -36,5 +36,8 @@
     "react-markdown": "^4.1.0",
     "react-typed": "^1.1.2",
     "reactstrap": "^8.0.0"
+  },
+  "devDependencies": {
+    "webpack": "^4.35.3"
   }
 }
