+++ web/static/admin/config.yml
@@ -1,12 +1,13 @@
 backend:
   name: github
   repo: mattseemon/mattseemon.in
+  branch: master_v2
 publish_mode: editorial_workflow
-media_folder: static/assets/images
+media_folder: web/static/assets/images
 collections:
   - name: "pages"
     label: "Pages"
-    folder: "posts"
+    folder: "web/content"
     create: true
     slug: "{{slug}}"
     fields:
@@ -17,7 +18,7 @@ collections:
       - { label: "Body", name: "body", widget: "markdown"}
   - name: "posts"
     label: "Blog Posts"
-    folder: "content/posts"
+    folder: "web/content/posts"
     create: true
     slug: "{{slug}}"
     fields:
@@ -27,7 +28,7 @@ collections:
       - { label: "Body", name: "body", widget: "markdown"}
   - name: "writing"
     label: "Writing"
-    folder: "content/writing"
+    folder: "web/content/articles"
     create: true
     slug: "{{slug}}"
     fields:
