+++ web/static/admin/index.html
@@ -3,10 +3,13 @@
 <head>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>mattseemon.in - Content Manager</title>
+  <title>Matt Seemon - Content Manager</title>
 </head>
 <body>
   <!-- Include the script that builds the page and powers Netlify CMS -->
-  <script src="https://unpkg.com/netlify-cms@^2.0.0/dist/netlify-cms.js"></script>
+  <script src="https://unpkg.com/netlify-cms@^2.9.6/dist/netlify-cms.js"></script>
+  <script>
+    CMS.registerPreviewStyle("../assets/css/main.css");
+  </script>
 </body>
 </html>
